PC2/Python/requirements.txt
paho-mqtt>=2.0
pymongo>=4.0
mysql-connector-python>=8.0
